+++ examples/pong_tutorial_03/pong.rs
@@ -76,7 +76,7 @@ fn load_sprite_sheet(world: &mut World) -> SpriteSheetHandle {
         loader.load(
             "texture/pong_spritesheet.png",
             PngFormat,
-            TextureMetadata::srgb(),
+            TextureMetadata::srgb_scale(),
             (),
             &texture_storage,
         )